Backend/requirements.txt
#basic
fastapi
uvicorn
pydantic_settings
pydantic
sqlalchemy
asyncpg
alembic
psycopg2-binary
datetime
passlib
python-jose[cryptography]
google-auth
aiosmtplib